requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
freezegun==1.1.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
uWSGI==2.0.19.1
